move anndata ext functions mostly relevant for testing to test
